SDC-3734. S3 origin error handling should allow copy not move
Reviewed-on: https://review.streamsets.net/4188

+++ aws-lib/src/main/java/com/streamsets/pipeline/stage/origin/s3/AmazonS3Util.java
@@ -134,16 +134,19 @@ public class AmazonS3Util {
     return isEligible;
   }
 
-  static void move(
+  static void copy(
       AmazonS3Client s3Client,
       String srcBucket,
       String sourceKey,
       String destBucket,
-      String destKey
+      String destKey,
+      boolean isMove
   ) {
     CopyObjectRequest cp = new CopyObjectRequest(srcBucket, sourceKey, destBucket, destKey);
     s3Client.copyObject(cp);
-    s3Client.deleteObject(new DeleteObjectRequest(srcBucket, sourceKey));
+    if (isMove) {
+      s3Client.deleteObject(new DeleteObjectRequest(srcBucket, sourceKey));
+    }
   }
 
   static S3Object getObject(AmazonS3Client s3Client, String bucket, String objectKey) {
